add authentication check for manipulating listings

+++ router/listing.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const router = express.Router({mergeParams:true});
+const router = express.Router({ mergeParams: true });
 const flash = require("connect-flash");
 
 const Listing = require("../models/lists");
@@ -7,36 +7,43 @@ const ExpressError = require("../utils/ExpressError");
 const wrapAsync = require("../utils/wrapAsync");
 const validateListing = require("../utils/validateListing");
 
-
-
-
 // show listings
 router.get("/", wrapAsync(async (req, res) => {
     let lists = await Listing.find();
-    res.render("./listings/home.ejs", { lists});
+    res.render("./listings/home.ejs", { lists });
 }));
 
 //create Listing
 router.get("/new", (req, res) => {
-    res.render("./listings/form.ejs");
+    if (!req.isAuthenticated()) {
+        req.flash("error", "Login first to create listing");
+        res.redirect("/signin");
+    } else {
+        res.render("./listings/form.ejs");
+    }
 });
 router.post("/", validateListing, wrapAsync(async (req, res) => {
     const newListing = new Listing(req.body.listing);
     await newListing.save();
     req.flash("success", "New Listing Added");
-    res.redirect("/");
+    res.redirect("/listing");
 }));
 
 //edit Listing
 router.get("/:id/edit", wrapAsync(async (req, res) => {
-    let { id } = req.params;
-    let listing = await Listing.findById(id);
-
-    if (!listing) {
-        req.flash("error", "Listing dose not exist");
-        res.redirect("/");
+    if (!req.isAuthenticated()) {
+        req.flash("error", "Login first to edit listing");
+        res.redirect("/signin");
     } else {
-        res.render("./listings/update.ejs", { listing });
+        let { id } = req.params;
+        let listing = await Listing.findById(id);
+
+        if (!listing) {
+            req.flash("error", "Listing dose not exist");
+            res.redirect("/");
+        } else {
+            res.render("./listings/update.ejs", { listing });
+        }
     }
 }));
 router.patch("/:id", validateListing, wrapAsync(async (req, res) => {
@@ -57,19 +64,22 @@ router.patch("/:id", validateListing, wrapAsync(async (req, res) => {
 
 }));
 
-
 // delete Listing
 router.delete("/:id", wrapAsync(async (req, res) => {
-    let { id } = req.params;
-    const deleted = await Listing.findByIdAndDelete(id);
+    if (!req.isAuthenticated()) {
+        req.flash("error", "Login first to delete Listing");
+        res.redirect("/signin");
+    } else {
+        let { id } = req.params;
+        const deleted = await Listing.findByIdAndDelete(id);
 
-    if (!deleted) {
-        req.flash("error", "Listing dose not deleted");
-        res.redirect(`/listing/${id}`);
-    }
-    else {
-        req.flash("success", "Listing is deleted");
-        res.redirect("/listing");
+        if (!deleted) {
+            req.flash("error", "Listing dose not deleted");
+            res.redirect(`/listing/${id}`);
+        } else {
+            req.flash("success", "Listing is deleted");
+            res.redirect("/listing");
+        }
     }
 }));
